Improved network timeouts, reformatted.
The ClientReader now uses a class called NetworkTImer, which keeps track of the number of chunks downloaded in a batch, and if that number doesn't change for a default period of 10 seconds, the timer causes the ClientReader (which is executing wrangleChunks) to stop waiting, and either try to contact other ChunkServers, or move onto the next batch. This mechanism replaces the CountDownLatch used previously, whose downside was a pre-set timeout, static to network speed. Now, as long as chunks stream in, at a slowest rate of one every ten seconds, the download will not stop. Also, annoyingly, I reformatted the project to remove leading and trailing spaces inside if statements, while loops, etc. This will make the contributions page meaningless, unless there is some way to ignore whitespace changes.