File access, repair event, TCP changes.
Poring over the thread dumps of the various nodes taken during deadlock, I discovered that the deadlock was caused by a cascade of blocking calls to TCPSender's sendData() method. It would only happen after I'd deleted all the files at a particular server, and then attempted to read a large file. The server with the deleted files would send a massive burst of corruption reports to the Controller, which would fill up the TCP send buffer, and block until the Controller had the chance to start reading some of them. UDP wouldn't have had this issue, but there would have been a silent dropping of packets on the Controller end.

Many of the functions at the Controller are synchronized, and if any call blocks while holding that object monitor, requests stop being processed. That appears to have happened. Calls to the TCPSender sendData() method would block inside the Controller while the caller was holding the Controller's object monitor, preventing the Controller from churning through new events until the message could be sent.

But, the receiver of the Controller's sendData() call couldn't work through its events because it too was blocked on a send to another server. And so on, until the culprit was the server who sent all those corruption events to the Controller, who couldn't work through its events because it was blocked on the Controller.

For a long time, I thought the problem was with more traditional lock contention -- a misordering of lock acquisitions, either at the Controller or at the ChunkServer. If at the ChunkServer, perhaps involving access to files, which wasn't coded correctly anyway.

To fix the issue, I converted the TCPSender into a TCPSenderThread, which contains a blocking queue that other threads can add to. This allows the Controller to queue messages to send while it holds locks which stop it from proceeding, solving the Producer/Consumer problem. Incidentally, I discovered this issue back when I took the undergraduate version of this course, but had forgotten about the consequences until now, which is depressing.

I replaced the Chunk/ShardReader and Chunk/Shard/Writers with a Chunk/ShardProcessor and another data structure called the FileStreamer. This was only possible because I modified the FileMap to store both a FileMetadata and a Reentrant lock. Every time a value is retrieved from the FileMap, the lock for that record is also acquired, meaning no other thread has access to the mapped value. The holder can modify both the FileMetadata, and read the file without having to worry about concurrent access.

It feels a bit messy, but so be it. It is only made possible by the ConcurrentHashMap's compute() method, which allows a thread to access a single mapped value exclusively.

The TCPConnectionCache now uses a ConcurrentHashMap, which is what my design evolved into being from the P2P project. It doesn't use the data structure how it was intended to be used, but I think it does the trick just fine.

Now, the RepairShard/Chunk messages shuffle the route that the message should take, not just relying on the order of the servers[] array stored in the Controller's fileTable. This hopefully will load-balance the messaging better.

The FileSynchronizer was renamed to FileUtilities, as it is no longer relied upon to provide any synchronization -- that is now done by the FileMap.

A problem that persists is how the servers repair corrupt or missing files. Right now, if replicating, the Client asks the first server for the chunk, waits for timeout, then asks the second, and so on. If all three servers hold corrupt versions of the file, the read will fail, even if the parts of the files that are corrupt are non-overlapping. That is because the first server will deny the retrieve request, then notify the Controller of corruption, which will only repair its version of the chunk, not the versions of the chunk on other servers. Then the next server will deny the Client request, and so on. The read will fail, but hopefully the file will be repaired at each server.

Instead, either the repair request should attempt to repair all replications of the chunk, or the Client should request the same server multiple times for the same chunk, allowing time for the chunk to be repaired between requests. I don't know which solution is best.

Ideally, the Client should sent one request message per chunk, and that request message should traverse all the servers that could be used to serve that chunk until an uncorrupt chunk can be assembled. The chunk can then be served to the Client. Along the way, servers can use that message to fix their own copies, if possible. If not, they can notify the Controller of corruption. This wouldn't solve the issue of checking for corruption at all servers that hold copies of a file every time a request for that file is made. But, maybe that would be wasteful anyway.

Finally, removed the TCPConnection associated with each server from the ServerConnection the Controller uses to keep track of registrants. Using a TCPConnectionCache to communicate with registered servers is better practice anyway, because broken connections replaced automatically.