Download patch for replication schema
When a Client requests a chunk from a server, and the chunk is corrupt, the server denies the request. If all replications of the chunk are corrupt, the Client will be unable to download the file, which is fine.

But, if the three replications are individually corrupt, but can be combined to create an uncorrupted chunk, the Client will still be unable to read the file -- it will be denied by all three servers.

To solve this problem without changing the nature of the file requests/repairs, I just made the Client contact each server twice, the second time, hopefully, the chunk has been repaired. This is a band-aid fix, a better fix would be to change the way the file requests work, so that the Client only has to send out one request per chunk, and if the chunk is corrupt, the request is forwarded to another server by the first server, and so on. Repairs could happen along the way too to increase efficiency.