Consolidated chunk/shard messages, serve requests repair files automatically.
I wanted to consolidate/simplify the mechanisms for dealing with chunks and dealing with shards. There is now one message type for Requesting and Repairing chunks.

Chunk and ShardProcessors are now each FileProcessors, and which is used is abstracted based on the type of request/repair being dealt with. I think the code is more human readable. Plus separate functions were able to be combined.

I wasn't happy with the previous commit where, when the DFS is in replication mode, the Client could contact all three servers holding replications, and not get a valid chunk back, even though, between the three replications, there was a way to frankenstein a valid chunk. I got around this problem by making the Client contact a max of six servers -- contacting the three replication servers one at a time, two times in a row. This allowed for the servers to repair their chunks, and once repaired, the Client could, one at a time, request the chunk again. It worked, but it felt like a hack.

Instead, I changed the RequestChunk message so that, if the first server doesn't have a valid copy of the chunk, it attaches its valid slices to the message and passes the message along. This happens until the chunk can be successfully assembled, then it is served to the Client. If it can't be assembled, the request is denied.

After a complete chunk can be assembled, that server sends out non-corrupt versions of the chunk to servers that, along the route of the relay, indicated that their local chunk was corrupt.

This works well for chunks, because there are only three replications, and it is unlikely that the message will traverse all three servers.

What I didn't think about was how the technique would work with erasure coding. I just wasn't paying attention. The problem is that the message must traverse a minimum of six servers to reassemble the chunk. This makes for a total network usage of around 230KB transferred just to recover one chunk if erasure coding is being used. I love the idea of the Client sending a request to one server, and getting served the requested chunk from somewhere in the network, but this is probably too big a price to pay.

Will have to think about a way to satisfy both design elements -- the Client only having to contact one server, and the contents of the message not having to be relayed. Probably would be best to have the contacted server act as the getter for the other fragments, which could then be assembled and sent back to the Client. This server could detect corruptions, and use its fully assembled chunk to redistribute uncorrupted versions.