Refactored Client, added ls command.
Gave the Client a bit of a refactor. Not perfect, but more readable, a couple of smaller functions.

Client can now take and 'ls' command, listing the contents of the working directory. I just took the same function I wrote for the P2P StoreData node.

Added functionality for the 'put', 'get', and 'delete' commands.

The put command can now take a directory, and attempt to store all non-folder files in that directory to the DFS (just one folder deep). It can take multiple files/directories in a single command.

The get and put commands still can take multiple arguments, but can now take ranges. Instead of having to list out every number you want to get/delete, you can use "0..5" to represent a compression of "0 1 2 3 4 5". Will make it easier to work with lots of files at at time.